better NutritionScreen

+++ src/main/java/com/teammoeg/frostedheart/util/client/FGuis.java
@@ -88,10 +88,10 @@ public class FGuis {
      * @param smooth 平滑度(抗锯齿)(范围0-1f)(不要给太大的数值)
      */
     public static void drawRing(GuiGraphics guiGraphics, int x, int y, float innerRadius, float outerRadius,float startAngle,float endAngle, int innerColor,int outerColor,float smooth) {
-        float x2 = (int) (x + outerRadius);
-        float y2 = (int) (y + outerRadius);
-        float x1 = (int) (x - outerRadius);
-        float y1 = (int) (y - outerRadius);
+        float x2 =  (x + outerRadius);
+        float y2 =  (y + outerRadius);
+        float x1 =  (x - outerRadius);
+        float y1 =  (y - outerRadius);
 
 
         RenderSystem.setShader(FHShaders::getRingShader);
